Ajout de tests sur les KeyStore.

+++ KeyStore/FileKeyStore.php
@@ -3,7 +3,11 @@
 namespace Dedipanel\PHPSeclibWrapperBundle\KeyStore;
 
 use Dedipanel\PHPSeclibWrapperBundle\KeyStore\KeyStoreInterface;
+use Dedipanel\PHPSeclibWrapperBundle\KeyStore\Exception\KeyStoreInitializationException;
+use Dedipanel\PHPSeclibWrapperBundle\KeyStore\Exception\KeyStoreNotInitializedException;
+use Dedipanel\PHPSeclibWrapperBundle\KeyStore\Exception\KeyAlreadyExistsException;
 use Dedipanel\PHPSeclibWrapperBundle\KeyStore\Exception\KeyNotExistsException;
+use Dedipanel\PHPSeclibWrapperBundle\KeyStore\Exception\AlreadyInitializedKeyStore;
 
 /**
  * @author Albin Kerouanton
@@ -12,50 +16,120 @@ use Dedipanel\PHPSeclibWrapperBundle\KeyStore\Exception\KeyNotExistsException;
  */
 class FileKeyStore implements KeyStoreInterface
 {
-    private $dir;
+    private $storePath;
+    private $initialized = false;
 
     /**
-     * @param string $dir Directory absolute path (without trailing slash) containg keys
+     * @param string $storePath Absolute path (without trailing slash) containg keys
      */
-    public function __construct($dir)
+    public function __construct($storePath)
     {
-        $this->dir = $dir;
+        $this->storePath = rtrim($storePath, '/');
+        $this->initialized = false;
+    }
+    
+    /**
+     * {@inheritdoc}
+     */
+    public function initialize()
+    {
+        if (!file_exists($this->storePath) && @mkdir($this->storePath, 0775, true) !== true) {
+            throw new KeyStoreInitializationException();
+        }
+        
+        $this->initialized = true;
+        
+        return $this;
+    }
+    
+    /**
+     * {@inheritdoc}
+     */
+    public function isInitialized()
+    {
+        return $this->initialized;
     }
 
     /**
-     * @{inheritDoc}
+     * {@inheritdoc}
      */
     public function store($name, $content)
     {
+        if (!$this->isInitialized()) {
+            throw new KeyStoreNotInitializedException();
+        }
+        
+        $filepath = $this->getFilepath($name);
+        
+        if (file_exists($filepath)) {
+            throw new KeyAlreadyExistsException();
+        }
+        
         return file_put_contents($this->getFilepath($name), $content) !== false;
     }
 
     /**
-     * @{inheritDoc}
+     * {@inheritdoc}
      */
     public function retrieve($name)
     {
+        if (!$this->isInitialized()) {
+            throw new KeyStoreNotInitializedException();
+        }
+        
         $filepath = $this->getFilepath($name);
 
         if (!file_exists($filepath)) {
-            throw new KeyNotExistsException;
+            throw new KeyNotExistsException();
         }
 
         return file_get_contents($filepath);
     }
 
     /**
-     * @{inheritDoc}
+     * {@inheritdoc}
      */
     public function remove($name)
     {
+        if (!$this->isInitialized()) {
+            throw new KeyStoreNotInitializedException();
+        }
+        
         $filepath = $this->getFilepath($name);
 
         if (!file_exists($filepath)) {
-            throw new KeyNotExistsException;
+            throw new KeyNotExistsException();
         }
 
         return unlink($filepath);
+    }
+    
+    /**
+     * Set the store path (need to be done before initialization)
+     * 
+     * @param string $storePath
+     * 
+     * @return \Dedipanel\PHPSeclibWrapperBundle\KeyStore\FileKeyStore
+     */
+    public function setStorePath($storePath)
+    {
+        if ($this->isInitialized()) {
+            throw new AlreadyInitializedKeyStore();
+        }
+        
+        $this->storePath = $storePath;
+        
+        return $this;
+    }
+    
+    /**
+     * Get the store path
+     * 
+     * @return string
+     */
+    public function getStorePath()
+    {
+        return $this->storePath;
     }
 
     /**
@@ -67,6 +141,6 @@ class FileKeyStore implements KeyStoreInterface
      */
     private function getFilepath($name)
     {
-        return $this->dir . '/' . $name . '.key';
+        return $this->storePath . '/' . $name . '.key';
     }
 }
